fix filenames for edit_device; move pages out of single item folders
