+++ lib/widgets/pages/memberList/member_list_page.dart
@@ -13,8 +13,8 @@ import 'package:weforza/repository/settings_repository.dart';
 import 'package:weforza/widgets/common/generic_error.dart';
 import 'package:weforza/widgets/common/rider_search_filter_empty.dart';
 import 'package:weforza/widgets/pages/add_member/add_member_page.dart';
-import 'package:weforza/widgets/pages/exportMembers/exportMembersPage.dart';
-import 'package:weforza/widgets/pages/importMembers/importMembersPage.dart';
+import 'package:weforza/widgets/pages/export_members_page.dart';
+import 'package:weforza/widgets/pages/import_members_page.dart';
 import 'package:weforza/widgets/pages/memberDetails/member_details_page.dart';
 import 'package:weforza/widgets/pages/memberList/memberListEmpty.dart';
 import 'package:weforza/widgets/pages/memberList/memberListItem.dart';
