+++ lib/widgets/pages/rideList/rideListPage.dart
@@ -7,7 +7,7 @@ import 'package:weforza/model/ride.dart';
 import 'package:weforza/repository/ride_repository.dart';
 import 'package:weforza/widgets/common/generic_error.dart';
 import 'package:weforza/widgets/pages/add_ride/add_ride_page.dart';
-import 'package:weforza/widgets/pages/exportRides/exportRidesPage.dart';
+import 'package:weforza/widgets/pages/export_rides_page.dart';
 import 'package:weforza/widgets/pages/rideDetails/rideDetailsPage.dart';
 import 'package:weforza/widgets/pages/rideList/rideListEmpty.dart';
 import 'package:weforza/widgets/pages/rideList/rideListItem.dart';
